+++ misskey-websocket/src/broker/model.rs
@@ -5,12 +5,11 @@ use std::task::{Context, Poll};
 
 use crate::broker::channel::{ChannelPongSender, ResponseSender, ResponseStreamSender};
 use crate::error::Error;
-use crate::model::{ApiRequestId, ChannelId};
+use crate::model::{ApiRequestId, ChannelId, SubNoteId};
 
 use async_rwlock::RwLock;
 use futures::future::{BoxFuture, Future, FutureExt};
 use misskey_core::model::ApiResult;
-use misskey_core::streaming::SubNoteId;
 use serde_json::Value;
 use uuid::Uuid;