+++ misskey-core/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Blanket `impl`s to `Client`
+- `StreamingClient` for streaming connections
 - `UploadFileClient` for uploading files
 
 ### Changed